requirements.txt
requests
beautifulsoup4
pyyaml
lxml
html5lib
